Reformatted a bunch of code

+++ app/src/main/java/com/lukaswillsie/onlinechess/data/Game.java
@@ -8,7 +8,7 @@ import java.util.List;
 /**
  * A class that represents a chess game by encapsulating the information associated with it,
  * like the name of the game, the name of the user's opponent, whose turn it is, etc.
- *
+ * <p>
  * Objects of this kind should only be created AFTER A USER HAS BEEN LOGGED IN.
  */
 public class Game {
@@ -33,10 +33,10 @@ public class Game {
     /**
      * Initialize this Game object from the given source data. data is assumed to be a list of
      * Objects, all either Integers or Strings.
-     *
+     * <p>
      * See the Protocols.pdf document in the ChessServer repo for more details about how the server
      * sends game data to the app, but below is a summary of the details relevant to this class.
-     *
+     * <p>
      * After the app logs a user in, the server sends a sequence of batches of transmissions. Each
      * batch corresponds to a game the user is involved in, and consists of the same data points,
      * defined in the ServerData enum, always in the exact same order. The order is also defined in
@@ -44,15 +44,15 @@ public class Game {
      * to the "type" field in the ServerData enum, and then each batch of data points should be
      * converted into a list, keeping the same order as they were sent in by the server, and then
      * each list should be given to a different Game object via their initialize() method.
-     *
+     * <p>
      * This method will unpack the data from the server and extract the information relevant to
      * the app.
      *
      * @param serverData - a list of Strings and Integers corresponding to important data about a game,
-     *             sent by the server to the app
-     * @return  0 if this object is correctly initialized
-     *          1 if the data received by this object in the form of serverData is incorrect in
-     *          some way. In this event, the object logs the error before returning.
+     *                   sent by the server to the app
+     * @return 0 if this object is correctly initialized
+     * 1 if the data received by this object in the form of serverData is incorrect in
+     * some way. In this event, the object logs the error before returning.
      */
     public int initialize(List<Object> serverData) {
         this.data = new HashMap<>();
@@ -67,33 +67,31 @@ public class Game {
             String black = (String) this.getServerData(serverData, ServerData.BLACK);
 
             int state;
-            if(white.equals(username)) {
+            if (white.equals(username)) {
                 data.put(GameData.OPPONENT, black);
 
                 // The server sends the state of the game as a bit: 0 or 1, where 0 means that it's
                 // white's turn and 1 that it's black's turn. The state that we store in Game objects
                 // is 1 if it's the USER's turn, 0 otherwise.
-                state = (((Integer)this.getServerData(serverData, ServerData.STATE)).equals(0)) ? 1 : 0;
+                state = (this.getServerData(serverData, ServerData.STATE).equals(0)) ? 1 : 0;
                 data.put(GameData.STATE, state);
 
                 data.put(GameData.ARCHIVED, this.getServerData(serverData, ServerData.WHITE_ARCHIVED));
 
                 data.put(GameData.CHECK, this.getServerData(serverData, ServerData.WHITE_CHK));
-            }
-            else if(black.equals(username)) {
+            } else if (black.equals(username)) {
                 data.put(GameData.OPPONENT, white);
 
                 // The server sends the state of the game as a bit: 0 or 1, where 0 means that it's
                 // white's turn and 1 that it's black's turn. The state that we store in Game objects
                 // is 1 if it's the USER's turn, 0 otherwise.
-                state = (((Integer)this.getServerData(serverData, ServerData.STATE)).equals(1)) ? 1 : 0;
+                state = (this.getServerData(serverData, ServerData.STATE).equals(1)) ? 1 : 0;
                 data.put(GameData.STATE, state);
 
                 data.put(GameData.ARCHIVED, this.getServerData(serverData, ServerData.BLACK_ARCHIVED));
 
                 data.put(GameData.CHECK, this.getServerData(serverData, ServerData.BLACK_CHK));
-            }
-            else {
+            } else {
                 Log.e(tag, "Tried to create game " + this.data.get(GameData.GAMEID) + " but user is not a player in it");
                 return 1;
             }
@@ -103,18 +101,18 @@ public class Game {
             data.put(GameData.TURN, this.getServerData(serverData, ServerData.TURN));
 
             // The user has only been offered a draw if it's their turn AND a draw has been offered
-            data.put(GameData.DRAW_OFFERED, (state == 1 && ((Integer)this.getServerData(serverData, ServerData.DRAW_OFFERED)).equals(1)) ? 1 : 0);
+            data.put(GameData.DRAW_OFFERED, (state == 1 && this.getServerData(serverData, ServerData.DRAW_OFFERED).equals(1)) ? 1 : 0);
 
             data.put(GameData.DRAWN, this.getServerData(serverData, ServerData.DRAWN));
 
             // If the user won, set USER_WON and USER_LOST accordingly
-            if(this.getServerData(serverData, ServerData.WINNER).equals(username)) {
+            if (this.getServerData(serverData, ServerData.WINNER).equals(username)) {
                 data.put(GameData.USER_WON, 1);
                 data.put(GameData.USER_LOST, 0);
             }
             // If the user didn't win, and the WINNER field from ServerData is non-empty, then the
             // opponent won
-            else if(!this.getServerData(serverData, ServerData.WINNER).equals("")) {
+            else if (!this.getServerData(serverData, ServerData.WINNER).equals("")) {
                 data.put(GameData.USER_WON, 0);
                 data.put(GameData.USER_LOST, 1);
             }
@@ -125,16 +123,13 @@ public class Game {
             }
 
             return 0;
-        }
-        catch(ClassCastException e) {
+        } catch (ClassCastException e) {
             Log.e(tag, "Some data couldn't be correctly cast from list: " + serverData);
             return 1;
-        }
-        catch(NumberFormatException e) {
+        } catch (NumberFormatException e) {
             Log.e(tag, "Some data couldn't be parsed into int from list: " + serverData);
             return 1;
-        }
-        catch(IndexOutOfBoundsException e) {
+        } catch (IndexOutOfBoundsException e) {
             Log.e(tag, "There weren't enough data points in list: " + serverData);
             return 1;
         }
@@ -150,7 +145,7 @@ public class Game {
      * constructor for  details.
      *
      * @param serverData - the List to extract data from
-     * @param dataType - identifies which piece of data to extract
+     * @param dataType   - identifies which piece of data to extract
      * @return the given piece of data, returned as an Object, uncasted
      */
     private Object getServerData(List<Object> serverData, ServerData dataType) {
@@ -161,21 +156,19 @@ public class Game {
      * Mark this game as archived by the user.
      */
     public void setArchived(boolean archived) {
-        if(archived) {
+        if (archived) {
             this.data.put(GameData.ARCHIVED, 1);
-        }
-        else {
+        } else {
             this.data.put(GameData.ARCHIVED, 0);
         }
     }
 
     public String toString() {
-        if(this.data == null) {
+        if (this.data == null) {
             return "No game data";
-        }
-        else {
+        } else {
             StringBuilder builder = new StringBuilder();
-            for(GameData data : GameData.values()) {
+            for (GameData data : GameData.values()) {
                 builder.append(data.toString() + ": " + this.data.get(data).toString() + "\n");
             }
